requirements.txt

spotipy
google-generativeai
python-dotenv
